Промежуточный коммит, поменял логику сапроса размера корзины

+++ src/main/java/com/project/dao/abstraction/ShoppingCartDao.java
@@ -11,5 +11,5 @@ public interface ShoppingCartDao {
 
     List getMaxIdCartItem();
 
-    int getCartSize(Long cartId);
+    long getCartSize(Long cartId);
 }
